+++ packages/xo-web/src/common/intl/messages.js
@@ -184,6 +184,9 @@ const messages = {
   srSharedType: 'Shared {type}',
   srNotSharedType: 'Not shared {type}',
 
+  // ----- Home snapshots  -----
+  snapshotVmsName: 'Name',
+
   // ----- Common components -----
   sortedTableAllItemsSelected: 'All of them are selected',
   sortedTableNoItems: 'No items found',